+++ backend/main.py
@@ -1,11 +1,10 @@
 from aiohttp import web
-
-from app import create_app
+from backend.app import create_app
 
 app = create_app()
-web.run_app(app)
 
-
+if __name__ == '__main__':
+    web.run_app(app)
